Add custom Mutt mailcaps
Add default and graphical Mutt mailcaps to open PDF and HTML attachments
with my preferred terminal and graphical programs. Adds a macro for
opening an attachment with the graphical mailcap rather than the default
mailcap. Add pdftotext to the list of software to install on a new
system in README.md.