Remap Vim netrw gn to simply invoke :Ntree
The Vim netrw gn binding reports an error when selecting a regular file
rather than a directory to indicate target directory depth. While
behavior is correct aside from this, the error message appears in a
separate window and is annoying to close. This change simply remaps gn
to invoke the :Ntree command with no argument, which in this case has
the same behavior but does not report an "error."