dwm: Preserve floating win pos when switching mon
Updates my personal dwm configuration patch so that when applied it
causes dwm to store relative window position and re-apply this position
to floating windows when the window has changed monitors, both in cases
where the monitor configuration itself has changed and in cases where
the window has been moved between active monitors with different
geometries.

In effect, this resolves the issue in dwm where moving a floating window
between monitors causes the window to take on new and undesirable
position that does not correspond to the position of said windows on its
initial monitor.