xinitrc: Make wallpaper locally configurable
From here on out, wallpaper can be configured on a system-by-system
basis by placing (or symbolic linking) one or more files (as appropriate
for the number of displays in use) into the ~/.config/wallpaper.local
directory. These files must be PNGs. If they are not present, feh will
silently fail to set a desktop background.

Wallpapers can still be added/tracked in the ~/.config/wallpaper
directory, but the actual path to the specific wallpaper in use will no
longer be hardcoded and universal across all systems.