xdg-open: Exec programs to avoid race condition
In several cases (such as when xdg-open is invoked via Mutt to display
an attachment), the lifetime of a temporary file passed to xdg-open is
the lifetime of the xdg-open process itself, leading to a race condition
when the xdg-open script terminates before the file passed as an
argument has been dispatched to the appropriate application to be
"opened." The easiest way to avoid this is to simply exec programs
rather than running them in the background and then returning success.

It is worth noting that this may not match the behavior of the actual
xdg-open script itself.

For reference, see https://bugzilla.redhat.com/show_bug.cgi?id=653249.